Improve typing
