+++ src/config/categories/books/index.ts
@@ -1,9 +1,9 @@
-import { Category } from '../index';
+import { Category, CategoryEnum } from '../types';
 import colors from 'vuetify/src/util/colors';
 import { mdiBook } from '@mdi/js';
 
 export const category: Category = {
-  id: 'books',
+  id: CategoryEnum.books,
   icon: mdiBook,
   color: colors.cyan.darken2,
   position: 'right',
