+++ src/config/categories/games/index.ts
@@ -1,9 +1,9 @@
-import { Category } from '../index';
+import { Category, CategoryEnum } from '../types';
 import colors from 'vuetify/src/util/colors';
 import { mdiGamepad } from '@mdi/js';
 
 const category: Category = {
-  id: 'games',
+  id: CategoryEnum.games,
   icon: mdiGamepad,
   color: colors.lightGreen.darken2,
   position: 'left',
