+++ src/config/categories/movies/index.ts
@@ -1,9 +1,9 @@
-import { Category } from '../index';
+import { Category, CategoryEnum } from '../types';
 import colors from 'vuetify/src/util/colors';
 import { mdiMovie } from '@mdi/js';
 
 const category: Category = {
-  id: 'movies',
+  id: CategoryEnum.movies,
   icon: mdiMovie,
   color: colors.red.base,
   position: 'top',
