+++ src/config/categories/tv/index.ts
@@ -1,9 +1,9 @@
-import { Category } from '../index';
+import { Category, CategoryEnum } from '../types';
 import colors from 'vuetify/src/util/colors';
 import { mdiTelevision } from '@mdi/js';
 
 const category: Category = {
-  id: 'tv',
+  id: CategoryEnum.tv,
   icon: mdiTelevision,
   color: colors.orange.darken3,
   position: 'bottom',
