+++ src/config/sections/index.ts
@@ -1,7 +1,7 @@
 import { CATEGORY__HOME, CATEGORY__ARTICLES } from '@/router/constants';
 import CategoryHome from '@/views/CategoryHome.vue';
 import CategoryArticles from '@/views/CategoryArticles.vue';
-import categories from '@/config/categories';
+import { CategoryEnum } from '@/config/categories/types';
 import { mdiHome, mdiNewspaperVariantMultiple } from '@mdi/js';
 
 export default [
@@ -9,7 +9,7 @@ export default [
     id: 'home',
     title: 'Home',
     icon: mdiHome,
-    categories: [categories.books, categories.games, categories.movies, categories.tv],
+    categories: [CategoryEnum.books, CategoryEnum.games, CategoryEnum.movies, CategoryEnum.tv],
     route: {
       path: '',
       name: CATEGORY__HOME,
@@ -20,7 +20,7 @@ export default [
     id: 'articles',
     title: 'Articles',
     icon: mdiNewspaperVariantMultiple,
-    categories: [categories.books, categories.games, categories.movies, categories.tv],
+    categories: [CategoryEnum.books, CategoryEnum.games, CategoryEnum.movies, CategoryEnum.tv],
     route: {
       path: 'articles',
       name: CATEGORY__ARTICLES,
