+++ src/router/index.ts
@@ -3,7 +3,7 @@ import VueRouter, { RouteConfig } from 'vue-router';
 import Home from '@/views/Home.vue';
 import Category from '@/views/Category.vue';
 import { HOME, CATEGORY__HOME } from './constants';
-import { CategoryTypesEnum } from '@/config/categories';
+import { CategoryEnum } from '@/config/categories/types';
 import sections from '@/config/sections';
 
 Vue.use(VueRouter);
@@ -19,7 +19,7 @@ const routes: RouteConfig[] = [
     component: Category,
     props: true,
     beforeEnter(to, from, next) {
-      if (to.params.categoryId in CategoryTypesEnum) {
+      if (to.params.categoryId in CategoryEnum) {
         next();
       } else {
         next({ name: HOME });
@@ -28,7 +28,7 @@ const routes: RouteConfig[] = [
     children: sections.map((section) => ({
       ...section.route,
       beforeEnter(to, from, next) {
-        if (section.categories.map((category) => category.id).includes(to.params.categoryId)) {
+        if (section.categories.includes(to.params.categoryId as CategoryEnum)) {
           next();
         } else {
           next({ name: CATEGORY__HOME, params: { categoryId: to.params.categoryId } });
